Change server's log file descriptors to FILE pointers
First and foremost, this was done to get the benefits of buffered IO.
Secondarily, by making this switch now (rather than waiting until
Jormungandr is more stable), I avoid having to investigate a bug I was
having. (If a server was opened then killed, any future servers opened
in the same directory wouldn't have visible output in the log
files. Come to think of it, the text was probably just overwriting the
start of the log files, rather than replacing the files entierly like I
was expecting)