Unconditionally initialize parseUInt's out parameter
When compiling with -O3, gcc is smart enough to recognize that "val" is
not always initialized. However, gcc is *NOT* smart enough to recognize
that when we use the variable on line 98 (and line 87?) it is guaranteed
to have been initialized.

btw, the reason that gcc detects the "problem" with -O3 but not with any
of the lower optimization levels is that when a pointer to a variable is
passed to a function, it assumes that the variable is initialized. When
we use -O3, however, parseUInt is inlined and gcc can immediately see
that val is not actually guaranteed to be initialized.