Push unrun jobs back on the job stack
Because the value of getJob was being stored between calls to runJobs,
the job that runJobs ran first was not necessarly the job that was not
necessarily the highest priority job. If the prior call ran out of
slots, then the current call would first run the job that was the
highest priority at the time of the prior call to runJobs instead of
whatever was currently the highest priority job.